template/requirements/misc.txt
#!/usr/bin/env -S pip install -r

lsp-tree-sitter[misc]
